Changed the grid data structure.
The grid is now calculated as a 4-dimensional array in compute_grid(). After this the uvs in insert_postNeuron() are appended into a 2-dimensional list of lists. In select_random() the correct 2d-array is selected (works only for post right now) and based on the weights multiple cells are selected. Then a random item from the masks-list is selected for each selected cell.
Seems to be a lot faster (2 secs for N = 5000)!