Fixed grid not using synapse coordinates correctly
The grid used to ignore the uv-coordinates generated for the synapse layer and just forward the pre neuron uv coordinate.

Now the grid properly returns synapse coordinates.

Also fixed grid not using its gridmask for synapses.

Also optimized some steps in grid.select_random() that could be calculated once when all neurons are inserted into the grid.

On the hippocampus model there are no more mapping errors, but runtime is now about 2-3 times slower that before, but still faster than in master. Some optimizations are still to be done.