Fixed mesh operations for Blender 2.77
In Blendr 2.77, the mesh operations have changed as follows:

obj.closest_point_on_mesh() now returns an additional boolean in
the first parameter, confirming if there is a closest point.

obj.ray_cast() also returns an additional boolean, and now needs
the starting point, a direction and a length as parameter,
instead of a start and end point.